Rename `async.py` to avoid collision with async keyword in Python 3.7.0.  Addresses zeekay/flask-uwsgi-websocket#67
